requirements.txt
easyocr>=1.7.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

